# server/src/app.ts
import cors from "cors";
import express from "express";
import type { NextFunction, Request, Response } from "express";
import albumRouter from "./router/album.js";
import anniversaryRouter from "./router/anniversary.js";
import coupleRouter from "./router/couple.js";
import { isHttpError } from "./errors.js";
import mediaRouter from "./router/media.js";
import partnerChatRouter from "./router/partnerChat.js";
import uploadRouter from "./router/upload.js";
import userRouter from "./router/user.js";
import userinfoRouter from "./router/userinfo.js";
import wishRouter from "./router/wish.js";

const app = express();

interface RequestParseError {
  type?: string;
  status?: number;
  statusCode?: number;
}

function isRequestParseError(error: unknown): error is RequestParseError {
  if (!error || typeof error !== "object") {
    return false;
  }

  return "type" in error && "status" in error;
}

app.use(cors());
app.use("/upload", express.raw({ type: "*/*", limit: "100mb" }));
app.use(express.json());

// 路由
app.use("/album", albumRouter);
app.use("/anniversaries", anniversaryRouter);
app.use("/couple-space", coupleRouter);
app.use("/media", mediaRouter);
app.use("/partner-chat", partnerChatRouter);
app.use("/upload", uploadRouter);
app.use("/user", userRouter);
app.use("/userinfo", userinfoRouter);
app.use("/wishes", wishRouter);

app.use((_req, res) => {
  res.status(404).json({ message: "route not found" });
});

app.use((err: unknown, _req: Request, res: Response, _next: NextFunction) => {
  void _next;

  if (isHttpError(err)) {
    res.status(err.statusCode).json({ message: err.message });
    return;
  }

  if (isRequestParseError(err) && err.type === "entity.parse.failed") {
    res.status(err.statusCode || err.status || 400).json({
      message: "invalid json payload",
    });
    return;
  }

  console.error(err);
  res.status(500).json({ message: "internal server error" });
});

export { app };