# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "InSync API",
    "description": "Formal HTTP API contract derived from the current server/ implementation.\n\nSource of truth: Express routers and serialize* functions.\n\nCompatibility: OpenAPI 3.0.3 (Apifox / Swagger friendly).\nCoupleSummary (inside /userinfo) and CoupleSpace (/couple-space) are different models.\nUpload success body is { key, url } without message. Lists are not paginated.\n",
    "version": "1.0.0",
    "contact": {
      "name": "InSync"
    }
  },
...
  "paths": {
...
    "/userinfo": {
      "get": {
        "summary": "获取用户资料",
        "deprecated": false,
        "description": "",
        "operationId": "getUserInfo",
        "tags": [
          "User"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfoResponse"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Invalid or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "summary": "更新用户资料",
        "deprecated": false,
        "description": "",
        "operationId": "updateUserInfo",
        "tags": [
          "User"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfoResponse"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Invalid or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "Profile columns unsupported",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/couple-space": {
      "get": {
        "summary": "获取情侣空间",
        "deprecated": false,
        "description": "",
        "operationId": "getCoupleSpace",
        "tags": [
          "Couple"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Couple space",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoupleSpaceResponse"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "Missing tables",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "summary": "更新恋爱纪念日",
        "deprecated": false,
        "description": "",
        "operationId": "updateCoupleSpace",
        "tags": [
          "Couple"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCoupleSpaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoupleSpaceResponse"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "Bound relationship not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/anniversaries": {
      "get": {
        "summary": "获取纪念日列表",
        "deprecated": false,
        "description": "按 remainingDays、nextOccurrenceDate 排序；未绑定关系时返回空数组（仍为 200）。",
        "operationId": "getAnniversaries",
        "tags": [
          "Anniversary"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "List sorted by remainingDays then nextOccurrenceDate.\nEmpty array when user is not bound (still 200).\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnniversaryListResponse"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "创建纪念日",
        "deprecated": false,
        "description": "",
        "operationId": "createAnniversary",
        "tags": [
          "Anniversary"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnniversaryPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnniversaryItemResponse"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            },
            "headers": {}
          },
          "409": {
            "description": "Bound couple relationship not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/wishes": {
      "get": {
        "summary": "获取心愿列表",
        "deprecated": false,
        "description": "",
        "operationId": "getWishes",
        "tags": [
          "Wish"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Wishes (no pagination)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WishListResponse"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "创建心愿",
        "deprecated": false,
        "description": "",
        "operationId": "createWish",
        "tags": [
          "Wish"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WishItemResponse"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/album/media": {
      "get": {
        "summary": "获取相册媒体",
        "deprecated": false,
        "description": "含 legacy wish_records.media_urls 合成项（id 可能为负数）；无分页。",
        "operationId": "getAlbumMedia",
        "tags": [
          "Album"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Media list (no pagination)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumMediaListResponse"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "创建相册媒体",
        "deprecated": false,
        "description": "",
        "operationId": "createAlbumMedia",
        "tags": [
          "Album"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAlbumMediaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumMediaItemResponse"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/upload/media": {
      "post": {
        "summary": "上传媒体文件",
        "deprecated": false,
        "description": "请求体为原始二进制；成功响应为 `{ key }`，无 message。对象读取 URL 由各自的授权资源接口签发。",
        "operationId": "uploadMedia",
        "tags": [
          "Upload"
        ],
        "parameters": [
          {
            "name": "folder",
            "in": "query",
            "description": "Storage folder segment. Not Zod-validated; clients use values like `album`.\n",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-file-name",
            "in": "header",
            "description": "Original file name (used for extension)",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            },
            "*/*": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Uploaded — body is `{ key }` without message",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadMediaResponse"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/partner-chat": {
      "get": {
        "summary": "伴侣聊天 WebSocket",
        "deprecated": false,
        "description": "Not a JSON REST endpoint. Clients open a WebSocket to\n`ws(s)://<host>/partner-chat?token=<jwt>` (or Bearer on upgrade).\n\nRequires an active bound couple relationship; otherwise upgrade is rejected\nwith HTTP 403.\n\nClient → server message schemas:\n- PartnerChatClientMessageText\n- PartnerChatClientMessageAudio (exactly one of `audioObjectKey` or legacy `audioUrl`)\n- PartnerChatClientRead\n\nNew object keys must belong to the authenticated sender under `interact/<userId>/`. The server does not echo private keys or signed URLs over WebSocket; clients fetch playback URLs from `GET /partner-chat/messages/{id}/audio-url`.\n\nServer → client message schemas:\n- PartnerChatServerReady\n- PartnerChatServerMessage\n- PartnerChatServerDelivery\n- PartnerChatServerReadReceipt\n- PartnerChatServerError\n\nMessage `id` values are **strings** (unlike HTTP resource ids).\n",
        "operationId": "partnerChatUpgrade",
        "tags": [
          "PartnerChat"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "JWT when not using Authorization header",
            "required": false,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "101": {
            "description": "Switching Protocols (WebSocket)",
            "headers": {}
          },
          "401": {
            "description": "Unauthorized upgrade rejection (plain text, not ErrorMessage JSON)",
            "headers": {}
          },
          "403": {
            "description": "No bound partner (plain text)",
            "headers": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
...
  "components": {
    "schemas": {
      "ErrorMessage": {
        "type": "object",
        "required": [
          "message"
        ],
        "additionalProperties": false,
        "properties": {
          "message": {
            "type": "string",
            "description": "Human-readable error message (no stable machine code yet)"
          }
        },
        "description": "Uniform HTTP error body used by `HttpError`, JSON parse failures,\nunknown routes, and unhandled 500s.\n"
      },
      "DateOnly": {
        "type": "string",
        "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
        "description": "Calendar date `YYYY-MM-DD` (no time zone)",
        "examples": [
          "2024-06-01"
        ]
      },
      "DateOnlyNullable": {
        "type": [
          "string",
          "null"
        ],
        "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
        "description": "Calendar date or null when unset"
      },
...
      "IsoDateTimeNullable": {
        "type": [
          "string",
          "null"
        ],
        "format": "date-time",
        "description": "ISO 8601 timestamp or null"
      },
...
      "PartnerSummary": {
        "type": "object",
        "required": [
          "id",
          "username",
          "nickname",
          "avatar"
        ],
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer"
          },
          "username": {
            "type": "string"
          },
          "nickname": {
            "type": [
              "string",
              "null"
            ]
          },
          "avatar": {
            "type": [
              "string",
              "null"
            ],
            "description": "Avatar URL or null (field name is `avatar`, not avatarUrl)"
          }
        },
        "description": "Partner card used inside CoupleSummary and CoupleSpace"
      },
...
      "User": {
        "type": "object",
        "required": [
          "id",
          "username",
          "nickname",
          "avatar",
          "signature",
          "birthday",
          "gender",
          "coupleStatus",
          "couple",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer"
          },
          "username": {
            "type": "string"
          },
          "nickname": {
            "type": [
              "string",
              "null"
            ]
          },
          "avatar": {
            "type": [
              "string",
              "null"
            ]
          },
          "signature": {
            "type": [
              "string",
              "null"
            ]
          },
          "birthday": {
            "$ref": "#/components/schemas/DateOnlyNullable"
          },
          "gender": {
            "type": [
              "string",
              "null"
            ],
            "description": "Returned by GET/PUT userinfo; not updatable via current PUT body"
          },
          "coupleStatus": {
            "type": [
              "string",
              "null"
            ],
            "description": "Latest couple_relationships.status for this user, or null.\nOverlaps conceptually with `couple.isBound` — do not collapse in contract.\n"
          },
          "couple": {
            "$ref": "#/components/schemas/CoupleSummary"
          },
          "createdAt": {
            "$ref": "#/components/schemas/IsoDateTimeNullable"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/IsoDateTimeNullable"
          }
        },
        "description": "Canonical user entity from `/userinfo` serializeUser.\n\nClient gaps (types only; runtime JSON still includes these fields):\n- Web AuthUser/UserProfile: often omits birthday, gender, coupleStatus, createdAt, updatedAt\n- App AuthUser: often omits nested `couple` while keeping coupleStatus\n"
      },
      "UserInfoResponse": {
        "type": "object",
        "required": [
          "message",
          "user"
        ],
        "additionalProperties": false,
        "properties": {
          "message": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "UpdateUserProfileRequest": {
        "type": "object",
        "required": [
          "nickname",
          "avatar",
          "signature",
          "birthday"
        ],
        "additionalProperties": false,
        "properties": {
          "nickname": {
            "type": "string",
            "maxLength": 30,
            "description": "Empty string is stored as null"
          },
          "avatar": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 2048,
            "description": "Empty string is stored as null"
          },
          "signature": {
            "type": "string",
            "maxLength": 200,
            "description": "Empty string is stored as null"
          },
          "birthday": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/DateOnly"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional valid YYYY-MM-DD date that cannot be in the future; null clears the stored birthday"
          }
        }
      },
...
      "CoupleSpace": {
        "type": "object",
        "required": [
          "isBound",
          "partner",
          "relationship",
          "daysInLove",
          "activeInvite"
        ],
        "additionalProperties": false,
        "properties": {
          "isBound": {
            "type": "boolean"
          },
          "partner": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PartnerSummary"
              }
            ],
            "type": "null"
          },
          "relationship": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CoupleRelationship"
              }
            ],
            "type": "null"
          },
          "daysInLove": {
            "type": [
              "integer",
              "null"
            ]
          },
          "activeInvite": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CoupleInvite"
              }
            ],
            "type": "null"
          }
        },
        "description": "Full couple-space payload from `/couple-space`.\nDistinct from `CoupleSummary` on User. Do not merge schemas.\n"
      },
      "CoupleSpaceResponse": {
        "type": "object",
        "required": [
          "message",
          "coupleSpace"
        ],
        "additionalProperties": false,
        "properties": {
          "message": {
            "type": "string"
          },
          "coupleSpace": {
            "$ref": "#/components/schemas/CoupleSpace"
          }
        }
      },
...
      "UpdateCoupleSpaceRequest": {
        "type": "object",
        "required": [
          "anniversaryDate"
        ],
        "additionalProperties": false,
        "properties": {
          "anniversaryDate": {
            "$ref": "#/components/schemas/DateOnlyNullable"
          }
        }
      },
      "AnniversaryType": {
        "type": "string",
        "enum": [
          "love",
          "birthday",
          "holiday",
          "custom"
        ]
      },
      "AnniversaryRepeatType": {
        "type": "string",
        "enum": [
          "none",
          "yearly"
        ]
      },
...
      "AnniversaryPayload": {
        "type": "object",
        "required": [
          "title",
          "type",
          "originalDate",
          "repeatType",
          "reminderDaysBefore"
        ],
        "additionalProperties": false,
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "type": {
            "$ref": "#/components/schemas/AnniversaryType"
          },
          "originalDate": {
            "$ref": "#/components/schemas/DateOnly",
            "description": "Real calendar date in YYYY-MM-DD format, year 1000–9999; invalid dates return 400"
          },
          "repeatType": {
            "$ref": "#/components/schemas/AnniversaryRepeatType"
          },
          "reminderDaysBefore": {
            "type": "integer",
            "minimum": 0,
            "maximum": 30
          }
        },
        "description": "Used for both create and update (full replacement, not partial)"
      },
      "AnniversaryListResponse": {
        "type": "object",
        "required": [
          "message",
          "anniversaries"
        ],
        "additionalProperties": false,
        "properties": {
          "message": {
            "type": "string"
          },
          "anniversaries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Anniversary"
            }
          }
        }
      },
      "AnniversaryItemResponse": {
        "type": "object",
        "required": [
          "message",
          "anniversary"
        ],
        "additionalProperties": false,
        "properties": {
          "message": {
            "type": "string"
          },
          "anniversary": {
            "$ref": "#/components/schemas/Anniversary"
          }
        }
      },
...
      "WishListResponse": {
        "type": "object",
        "required": [
          "message",
          "wishes"
        ],
        "additionalProperties": false,
        "properties": {
          "message": {
            "type": "string"
          },
          "wishes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Wish"
            }
          }
        }
      },
      "WishItemResponse": {
        "type": "object",
        "required": [
          "message",
          "wish"
        ],
        "additionalProperties": false,
        "properties": {
          "message": {
            "type": "string"
          },
          "wish": {
            "$ref": "#/components/schemas/Wish"
          }
        }
      },
      "CreateWishRequest": {
        "type": "object",
        "required": [
          "title",
          "targetDate"
        ],
        "additionalProperties": false,
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "default": ""
          },
          "cover": {
            "type": "string",
            "description": "Legacy URL cover, retained for backward compatibility; omitted for private object-key covers"
          },
          "coverObjectKey": {
            "type": "string",
            "maxLength": 2048,
            "description": "Private uploaded album media object key; must belong to the authenticated uploader"
          },
          "targetDate": {
            "$ref": "#/components/schemas/DateOnly",
            "description": "Valid calendar date from 1000-01-01 through 9999-12-31"
          },
          "locationName": {
            "type": "string",
            "maxLength": 100,
            "default": ""
          },
          "latitude": {
            "type": [
              "number",
              "null"
            ],
            "minimum": -90,
            "maximum": 90,
            "default": null
          },
          "longitude": {
            "type": [
              "number",
              "null"
            ],
            "minimum": -180,
            "maximum": 180,
            "default": null
          },
          "budgetAmount": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0,
            "maximum": 2147483647,
            "default": null
          }
        }
      },
...
      "AlbumMediaType": {
        "type": "string",
        "enum": [
          "image",
          "video"
        ]
      },
...
      "AlbumMediaListResponse": {
        "type": "object",
        "required": [
          "message",
          "media"
        ],
        "additionalProperties": false,
        "properties": {
          "message": {
            "type": "string"
          },
          "media": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlbumMedia"
            }
          }
        }
      },
      "AlbumMediaItemResponse": {
        "type": "object",
        "required": [
          "message",
          "media"
        ],
        "additionalProperties": false,
        "properties": {
          "message": {
            "type": "string"
          },
          "media": {
            "$ref": "#/components/schemas/AlbumMedia"
          }
        }
      },
      "CreateAlbumMediaRequest": {
        "type": "object",
        "required": [
          "mediaType",
          "objectKey"
        ],
        "additionalProperties": false,
        "properties": {
          "mediaType": {
            "$ref": "#/components/schemas/AlbumMediaType"
          },
          "objectKey": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2048,
            "description": "Object storage key returned by the authenticated upload endpoint"
          },
          "thumbnailUrl": {
            "type": "string",
            "default": ""
          },
          "takenAt": {
            "$ref": "#/components/schemas/DateOnly"
          },
          "locationName": {
            "type": "string",
            "maxLength": 100,
            "default": ""
          },
          "latitude": {
            "type": [
              "number",
              "null"
            ],
            "default": null
          },
          "longitude": {
            "type": [
              "number",
              "null"
            ],
            "default": null
          }
        }
      },
...
      "UploadMediaResponse": {
        "type": "object",
        "required": [
          "key"
        ],
        "additionalProperties": false,
        "properties": {
          "key": {
            "type": "string",
            "description": "Object storage key"
          }
        },
        "description": "Success body returns a private object key only; it has NO `url` or `message` field.\n"
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT from `POST /user/login`. Payload `sub` is the user id as a string.\nHeader: `Authorization: Bearer <token>`.\n"
      }
    },
...
  },
  "servers": [],
  "security": []
}